requirements.txt
plaid-python==34.0.0
python-dotenv==1.0.1